requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional LLM providers (uncomment as needed)
# For Ollama (local)
ollama>=0.0.1

# For OpenAI
# openai>=1.0.0
# langchain-openai>=0.0.1

# For Anthropic (Claude)
# anthropic>=0.7.0
# langchain-anthropic>=0.0.1

# For HuggingFace
# huggingface-hub>=0.16.0
# langchain-huggingface>=0.0.1

# Development
pytest>=7.0
black>=23.0
ruff>=0.0.200
